Explicitly configing log4j
Up to now the logstash has been logging to stdout, this was outside
our control and was filling up the disk. Added a log4j2 configuration
file, ripped from an example, when brings the logs under our explicit
control.

This config rolls daily and at 500MB and keeps the last three files